third_party/WebKit/LayoutTests/svg/animations/getCurrentTime-pause-unpause-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
PASS successfullyParsed is true

TEST COMPLETE
PASS svg.getCurrentTime() is 0
PASS svg.getCurrentTime() is 0.05
PASS svg.getCurrentTime() is 0.05
